# docs/swagger.yaml
definitions:
...
  controllers.audioResponse:
    properties:
      filename:
        example: audio_1.m3u8
        type: string
      language:
        example: jpn
        type: string
    type: object
...
  controllers.errorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  controllers.mediaFileResponse:
    properties:
      audios:
        items:
          $ref: '#/definitions/controllers.audioResponse'
        type: array
      createdAt:
        example: "2023-05-07T20:31:28.327382+02:00"
        type: string
      duration:
        example: 1450.76
        type: number
      filename:
        example: index.m3u8
        type: string
      id:
        example: eec1d6b7-97c9-47e9-846b-6817d0e3d4ed
        type: string
      subtitles:
        items:
          $ref: '#/definitions/controllers.subtitleResponse'
        type: array
      updatedAt:
        example: "2023-05-07T20:31:28.327382+02:00"
        type: string
    type: object
...
  controllers.subtitleResponse:
    properties:
      filename:
        example: subtitle_1.vtt
        type: string
      language:
        example: fre
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    This is the API for the Media Service application
    This help to give info about the media files and metadata
    This also help to manage the media files for admins
  title: Media Service API
paths:
...
  /media-file/file-tmdb/{id}:
    get:
      description: Get media file info by its Media TMDB ID
      parameters:
      - description: TMDB ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.mediaFileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.errorResponse'
      summary: Get media file info by its Media TMDB ID
      tags:
      - Media File
...
swagger: "2.0"